Switched to double type for Float

+++ inc/ast.h
@@ -75,7 +75,7 @@ public:
 
   union {
     int Int;
-    float Float;
+    double Float;
   } NumericData;
 
 protected:
